refactor(payments): extract loadStripeElements to a service
