+++ src/app/payment/views/IntegratedCheckoutView/CheckoutViewWrapper.tsx
@@ -1,13 +1,7 @@
 import { BaseSyntheticEvent, useCallback, useEffect, useReducer, useRef, useState } from 'react';
 import { Elements } from '@stripe/react-stripe-js';
 import { useSelector } from 'react-redux';
-import {
-  Stripe,
-  StripeElements,
-  StripeElementsOptions,
-  StripeElementsOptionsMode,
-  loadStripe,
-} from '@stripe/stripe-js';
+import { Stripe, StripeElements, StripeElementsOptions, StripeElementsOptionsMode } from '@stripe/stripe-js';
 import { UserSettings } from '@internxt/sdk/dist/shared/types/userSettings';
 import { DisplayPrice, UserType } from '@internxt/sdk/dist/drive/payments/types';
 
@@ -33,7 +27,7 @@ import { useThemeContext } from '../../../theme/ThemeProvider';
 import { getProductAmount } from '../../components/checkout/ProductCardComponent';
 import authCheckoutService from '../../services/auth-checkout.service';
 import { checkoutReducer, initialStateForCheckout } from '../../store/checkoutReducer';
-import { AuthMethodTypes, CouponCodeData, PlanData, RequestedPlanData } from '../../types';
+import { AuthMethodTypes, CouponCodeData, RequestedPlanData } from '../../types';
 import CheckoutView from './CheckoutView';
 import ChangePlanDialog from '../../../newSettings/Sections/Account/Plans/components/ChangePlanDialog';
 import { fetchPlanPrices, getStripe } from '../../../newSettings/Sections/Account/Plans/api/plansApi';
@@ -94,11 +88,6 @@ const RETURN_URL_DOMAIN = IS_PRODUCTION ? process.env.REACT_APP_HOSTNAME : 'http
 
 let stripe;
 
-export const stripePromise = (async () => {
-  const stripeKey = IS_PRODUCTION ? process.env.REACT_APP_STRIPE_PK : process.env.REACT_APP_STRIPE_TEST_PK;
-  return await loadStripe(stripeKey);
-})();
-
 function savePaymentDataInLocalStorage(
   subscriptionId: string | undefined,
   paymentIntentId: string | undefined,
@@ -120,13 +109,13 @@ const CheckoutViewWrapper = () => {
   const { checkoutTheme } = useThemeContext();
   const [state, dispatchReducer] = useReducer(checkoutReducer, initialStateForCheckout);
   const isAuthenticated = useAppSelector((state) => state.user.isAuthenticated);
-  const user = useSelector<RootState, UserSettings>((state) => state.user.user!);
+  const user = useSelector<RootState, UserSettings | undefined>((state) => state.user.user);
   const { doRegister } = useSignUp('activate');
   const userAuthComponentRef = useRef<HTMLDivElement>(null);
 
   const fullName = `${user?.name} ${user?.lastname}`;
   const isUserAuthenticated = !!user;
-  const isAnyError = state.error?.coupon || state.error?.auth || state.error?.stripe;
+  const thereIsAnyError = state.error?.coupon || state.error?.auth || state.error?.stripe;
 
   const {
     onRemoveAppliedCouponCode,
@@ -142,12 +131,12 @@ const CheckoutViewWrapper = () => {
     setUserNameFromElementAddress,
     setSeatsForBusinessSubscription,
   } = useCheckout(dispatchReducer);
+
   const [isUpsellSwitchActivated, setIsUpsellSwitchActivated] = useState<boolean>(false);
   const [isCheckoutReadyToRender, setIsCheckoutReadyToRender] = useState<boolean>(false);
   const [isUpdateSubscriptionDialogOpen, setIsUpdateSubscriptionDialogOpen] = useState<boolean>(false);
   const [isUpdatingSubscription, setIsUpdatingSubscription] = useState<boolean>(false);
-  const [individualPrices, setIndividualPrices] = useState<DisplayPrice[]>();
-  const [businessPrices, setBusinessPrices] = useState<DisplayPrice[]>();
+  const [prices, setPrices] = useState<DisplayPrice[]>();
   const [country, setCountry] = useState<string>('');
   const {
     authMethod,
@@ -160,13 +149,12 @@ const CheckoutViewWrapper = () => {
     promoCodeName,
     seatsForBusinessSubscription,
   } = state;
-  const prices = currentSelectedPlan?.type === UserType.Individual ? individualPrices : businessPrices;
   const canChangePlanDialogBeOpened = prices && currentSelectedPlan && isUpdateSubscriptionDialogOpen;
 
   const userInfo: UserInfoProps = {
     name: fullName,
     avatar: avatarBlob,
-    email: user?.email,
+    email: user?.email ?? '',
   };
 
   const upsellManager = {
@@ -201,8 +189,6 @@ const CheckoutViewWrapper = () => {
       return;
     }
 
-    fetchDataAndSetPrices();
-
     handleFetchSelectedPlan(planId, currencyValue)
       .then((plan) => {
         if (checkoutTheme && plan) {
@@ -213,9 +199,15 @@ const CheckoutViewWrapper = () => {
             });
           }
 
-          const { backgroundColor, textColor, borderColor, borderInputColor, labelTextColor } =
-            THEME_STYLES[checkoutTheme as string];
-          loadStripeElements(textColor, backgroundColor, borderColor, borderInputColor, labelTextColor, plan);
+          checkoutService.loadStripeElements(THEME_STYLES[checkoutTheme as string], onLoadElements, plan);
+          getStripe(stripe).then((stripePromise) => (stripe = stripePromise));
+          fetchPrices(plan.selectedPlan.type)
+            .then(setPrices)
+            .catch((error) => {
+              const errorCasted = errorService.castError(error);
+              errorService.reportError(errorCasted);
+            });
+          setIsCheckoutReadyToRender(true);
         }
       })
       .catch((error) => {
@@ -226,8 +218,6 @@ const CheckoutViewWrapper = () => {
           navigationService.push(AppView.Signup);
         }
       });
-
-    setIsCheckoutReadyToRender(true);
   }, [checkoutTheme]);
 
   useEffect(() => {
@@ -248,7 +238,7 @@ const CheckoutViewWrapper = () => {
   }, [promoCodeName]);
 
   useEffect(() => {
-    if (isAnyError) {
+    if (thereIsAnyError) {
       setTimeout(() => {
         setError('auth', undefined);
         setError('stripe', undefined);
@@ -257,18 +247,9 @@ const CheckoutViewWrapper = () => {
     }
   }, [state.error]);
 
-  const fetchDataAndSetPrices = useCallback(async () => {
-    try {
-      const individualPrices = await fetchPlanPrices(UserType.Individual);
-      const businessPrices = await fetchPlanPrices(UserType.Business);
-
-      setIndividualPrices(individualPrices);
-      setBusinessPrices(businessPrices);
-    } catch (error) {
-      const errorCasted = errorService.castError(error);
-      errorService.reportError(errorCasted);
-    }
-  }, []);
+  const onLoadElements = (stripeElementsOptions: StripeElementsOptions) => {
+    setStripeElementsOptions(stripeElementsOptions);
+  };
 
   const onChangePlanClicked = async (priceId: string, currency: string) => {
     setIsUpdatingSubscription(true);
@@ -329,6 +310,12 @@ const CheckoutViewWrapper = () => {
     }
   };
 
+  const fetchPrices = useCallback(async (userType: UserType) => {
+    const individualPrices = await fetchPlanPrices(userType);
+
+    return individualPrices;
+  }, []);
+
   const onCheckoutButtonClicked = async (
     formData: IFormValues,
     event: BaseSyntheticEvent<object, any, any> | undefined,
@@ -340,7 +327,7 @@ const CheckoutViewWrapper = () => {
     setIsUserPaying(true);
 
     const { email, password, companyName, companyVatId } = formData;
-    const userData = getUserInfo(user, email, userNameFromAddressElement, fullName);
+    const userData = getUserInfo(email, userNameFromAddressElement, fullName, user);
     const isStripeNotLoaded = !stripeSDK || !elements;
     const customerName = companyName ?? userData.name;
 
@@ -428,75 +415,6 @@ const CheckoutViewWrapper = () => {
     }
   };
 
-  const loadStripeElements = async (
-    textColor: string,
-    backgroundColor: string,
-    borderColor: string,
-    borderInputColor: string,
-    labelTextColor: string,
-    plan: PlanData,
-  ) => {
-    const stripeElementsOptions: StripeElementsOptions = {
-      appearance: {
-        labels: 'above',
-        variables: {
-          spacingAccordionItem: '8px',
-          colorPrimary: textColor,
-        },
-        theme: 'flat',
-        rules: {
-          '.AccordionItem:hover': {
-            color: textColor,
-          },
-          '.Block': {
-            backgroundColor: backgroundColor,
-          },
-          '.TermsText': {
-            color: textColor,
-          },
-          '.AccordionItem': {
-            borderRadius: '16px',
-            borderColor: borderColor,
-            border: '1px solid',
-            backgroundColor: backgroundColor,
-          },
-          '.Input': {
-            backgroundColor: `${backgroundColor}`,
-            borderRadius: '0.375rem',
-            color: textColor,
-            border: `1px solid ${borderInputColor}`,
-          },
-          '.Input:focus': {
-            backgroundColor: `${backgroundColor}`,
-            // borderColor: borderInputColor,
-            boxShadow: `0px 0px 4px ${BORDER_SHADOW}`,
-            border: `0.5px solid ${BORDER_SHADOW}`,
-          },
-          '.Input::selection': {
-            backgroundColor: `${backgroundColor}`,
-            // borderColor: borderInputColor,
-            border: `0.5px solid ${BORDER_SHADOW}`,
-          },
-          '.Label': {
-            color: labelTextColor,
-            fontSize: '0.875rem',
-          },
-          '.RedirectText': {
-            color: textColor,
-          },
-        },
-      },
-      mode: plan?.selectedPlan.interval === 'lifetime' ? 'payment' : 'subscription',
-      amount: plan?.selectedPlan.amount,
-      currency: plan?.selectedPlan.currency,
-      payment_method_types: ['card', 'paypal'],
-    };
-
-    setStripeElementsOptions(stripeElementsOptions);
-
-    stripe = await stripePromise;
-  };
-
   const handleFetchSelectedPlan = async (planId: string, currency?: string) => {
     const plan = await checkoutService.fetchPlanById(planId, currency);
     setPlan(plan);
@@ -523,10 +441,10 @@ const CheckoutViewWrapper = () => {
     await databaseService.clear();
     localStorageService.clear();
     RealtimeService.getInstance().stop();
-    setAuthMethod('signIn');
+    setAuthMethod('signUp');
   };
 
-  const getUserInfo = (user: UserSettings, email: string, userNameFromAddressElement: string, fullName: string) => {
+  const getUserInfo = (email: string, userNameFromAddressElement: string, fullName: string, user?: UserSettings) => {
     let userData;
 
     if (user) {
